Create Robot.java_03_17_17
I made a finishedMoving method that resets the encoders after every task/command to fix the overlap problem from the encoders.
The two cameras work!!! XDDDDD
I made the robot turn slower in autonomous, but I might just let it be within 1 degree? Is it okay to just have the robot spin a full circle once or twice? I'm thinking about how long it'll take to turn for the time constraints. 
I got rid of the while loops inside the moveForward and changeAngle methods. 
yay robots!